// packages/crusher-electron-app/src/extension/interfaces/topLevelAction.ts
export enum TOP_LEVEL_ACTION {
	TOGGLE_INSPECT_MODE = "TOGGLE_INSPECT_MODE",
	TAKE_PAGE_SCREENSHOT = "TAKE_PAGE_SCREENSHOT",
	WAIT = "WAIT",
	SHOW_SEO_MODAL = "SHOW_SEO_MODAL",
	CUSTOM_CODE = "CUSTOM_CODE",
	RUN_AFTER_TEST = "RUN_AFTER_TEST",
	SCROLL_AND_TAKE_SCREENSHOT = "SCROLL_AND_TAKE_SCREENSHOT",
	VERIFY_LINK = "VERIFY_LINK",
}